python/demo1/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
absl-py==2.0.0
    # via
    #   tensorboard
    #   tensorflow
astunparse==1.6.3
    # via tensorflow
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
beautifulsoup4==4.12.2
    # via google
cachetools==5.3.2
    # via google-auth
certifi==2023.11.17
    # via
    #   kfp-server-api
    #   kubernetes
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via kfp
cloudml-hypertune==0.1.0.dev6
    # via -r requirements.in
cloudpickle==3.0.0
    # via kfp
datetime==5.4
    # via -r requirements.in
deprecated==1.2.14
    # via kfp
docstring-parser==0.15
    # via kfp
fire==0.5.0
    # via
    #   -r requirements.in
    #   kfp
flatbuffers==23.5.26
    # via tensorflow
gast==0.4.0
    # via tensorflow
google==3.0.0
    # via -r requirements.in
google-api-core[grpc]==2.15.0
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-resource-manager
    #   google-cloud-storage
google-auth==2.25.2
    # via
    #   google-api-core
    #   google-auth-oauthlib
    #   google-cloud-core
    #   google-cloud-storage
    #   kfp
    #   kubernetes
    #   tensorboard
google-auth-oauthlib==0.4.6
    # via tensorboard
google-cloud-aiplatform==1.39.0
    # via
    #   -r requirements.in
    #   google-cloud-pipeline-components
google-cloud-bigquery==3.14.1
    # via google-cloud-aiplatform
google-cloud-core==2.4.1
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-pipeline-components==0.1.4
    # via -r requirements.in
google-cloud-resource-manager==1.11.0
    # via google-cloud-aiplatform
google-cloud-storage==2.14.0
    # via
    #   google-cloud-aiplatform
    #   kfp
google-crc32c==1.5.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-pasta==0.2.0
    # via tensorflow
google-resumable-media==2.7.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos[grpc]==1.62.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.13.0
    # via google-cloud-resource-manager
grpcio==1.60.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   tensorboard
    #   tensorflow
grpcio-status==1.48.2
    # via google-api-core
h5py==3.10.0
    # via tensorflow
idna==3.6
    # via requests
jsonschema==4.20.0
    # via kfp
jsonschema-specifications==2023.12.1
    # via jsonschema
keras==2.11.0
    # via tensorflow
kfp==1.4.0
    # via
    #   -r requirements.in
    #   google-cloud-pipeline-components
kfp-pipeline-spec==0.1.16
    # via kfp
kfp-server-api==1.8.5
    # via kfp
kubernetes==11.0.0
    # via kfp
libclang==16.0.6
    # via tensorflow
markdown==3.5.1
    # via tensorboard
markupsafe==2.1.3
    # via werkzeug
numpy==1.26.2
    # via
    #   h5py
    #   opt-einsum
    #   shapely
    #   tensorboard
    #   tensorflow
oauthlib==3.2.2
    # via requests-oauthlib
opt-einsum==3.3.0
    # via tensorflow
packaging==23.2
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   tensorflow
proto-plus==1.23.0
    # via
    #   google-cloud-aiplatform
    #   google-cloud-resource-manager
protobuf==3.19.6
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-resource-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   kfp-pipeline-spec
    #   proto-plus
    #   tensorboard
    #   tensorflow
pyasn1==0.5.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
python-dateutil==2.8.2
    # via
    #   google-cloud-bigquery
    #   kfp-server-api
    #   kubernetes
pytz==2023.3.post1
    # via datetime
pyyaml==6.0.1
    # via
    #   kfp
    #   kubernetes
referencing==0.32.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   kubernetes
    #   requests-oauthlib
    #   requests-toolbelt
    #   tensorboard
requests-oauthlib==1.3.1
    # via
    #   google-auth-oauthlib
    #   kubernetes
requests-toolbelt==1.0.0
    # via kfp
rpds-py==0.16.2
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via google-auth
shapely==2.0.2
    # via google-cloud-aiplatform
six==1.16.0
    # via
    #   astunparse
    #   fire
    #   google-pasta
    #   kfp-server-api
    #   kubernetes
    #   python-dateutil
    #   tensorflow
soupsieve==2.5
    # via beautifulsoup4
strip-hints==0.1.10
    # via kfp
tabulate==0.9.0
    # via kfp
tensorboard==2.11.2
    # via tensorflow
tensorboard-data-server==0.6.1
    # via tensorboard
tensorboard-plugin-wit==1.8.1
    # via tensorboard
tensorflow==2.11.0
    # via -r requirements.in
tensorflow-estimator==2.11.0
    # via tensorflow
tensorflow-io==0.31.0
    # via -r requirements.in
tensorflow-io-gcs-filesystem==0.31.0
    # via
    #   tensorflow
    #   tensorflow-io
termcolor==2.4.0
    # via
    #   fire
    #   tensorflow
typing-extensions==4.9.0
    # via tensorflow
urllib3==2.1.0
    # via
    #   kfp-server-api
    #   kubernetes
    #   requests
websocket-client==1.7.0
    # via kubernetes
werkzeug==3.0.1
    # via tensorboard
wheel==0.42.0
    # via
    #   astunparse
    #   strip-hints
    #   tensorboard
wrapt==1.14.1
    # via
    #   deprecated
    #   tensorflow
zope-interface==6.1
    # via datetime

# The following packages are considered to be unsafe in a requirements file:
# setuptools
